# Feature: Speed-up tests by caching client