Execute tests faster by only instantiating one client when using the `helper.getClient()` helper method.