@timothyej 2015-11-04
Question is; are we okay with losing isolation between tests for this performance improvement?